NODE_ENV is not recognised on Windows OS
NODE_ENV is not a recognised command on Windows, the work around is to use the cross-env package or to have two separate scripts, one for Windows and one for Unix based systems. I figured using the cross-env is much more friendly and will save confusion, tested on both Windows and Mac.